requirements.txt
beautifulsoup4==4.9.3
cloudscraper
requests
pyTelegramBotAPI
schedule
fcache
logtail-python